# api/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Gatte management API
  version: 1.6.0
  summary: The backend every Gatte front talks to (design/adr/0040).
  description: |
    Served by `mcp-gateway admin` over UNIX sockets only. There is no TCP
    listener. HTTP/1.1, JSON bodies, snake_case fields, times in RFC 3339
    UTC.

    Two sockets, two processes, two directories, two disjoint route sets.
    Every operation below carries `x-gatte-socket`:

    - `operator`: `/run/mcp-gateway-admin/operator/operator.sock`, served
      by a process running as the gateway's service account; the file is
      group `gatte-operators`, `0660`, in a directory of its own.
    - `accounts`: `/run/mcp-gateway-admin/accounts/accounts.sock`, served
      only by a root process; the file is `root:root 0600`, or
      `root:ACCOUNT_GROUP 0660` when `[admin] account_group` delegates
      accounts (both settings are required; the backend refuses a mismatch
      at startup). Its directory is `root:root 0755`.

    A route of the other socket answers 404 `wrong_socket` with
    `details.socket`; a route that does not exist answers 404
    `unknown_route`.

    ## Who serves

    `pkg/adminapi` reads the server's credentials after connecting and
    refuses an accounts socket whose server is not uid 0 (under systemd the
    kernel reports systemd, which created the socket). `WithServerUID` pins
    the operator socket's server.

    ## Who the operator is

...
    `GET /v1/whoami` returns what the backend decided.

...

...
servers:
...
components:
...
  responses:
    Error:
      description: |
        Any failure. `code` is stable and machine-readable; `message` is
        one English sentence for a person and may change.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /v1/overview:
    get:
      operationId: overview
      tags: [overview]
      x-gatte-socket: [operator]
      summary: What needs attention, the counts, and the latest refusals.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Overview'}
        default: {$ref: '#/components/responses/Error'}
  /v1/tools:
    get:
      operationId: listTools
      tags: [tools]
      x-gatte-socket: [operator]
      summary: The quarantine entries (`tool list`).
      parameters:
        - {name: server, in: query, schema: {type: string}, description: Only this backend.}
        - name: show
          in: query
          schema: {type: string, enum: [all, review, approved], default: all}
          description: '`review` is every entry not approved; `approved` the rest. `counts` is always over all.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolList'}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/tools/approve:
    post:
      operationId: approveTool
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: Approve the fingerprint that was reviewed, and only that one. Writes `(tool approve)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApproveRequest'}
      responses:
        '200':
          description: |
            Approved, or already approved at exactly this fingerprint (`changed`
            false). A sensitive tool (1.5.0) is approved and still `usable:
            false` until cleared with `clearTool`; the warning
            `sensitive_uncleared` says so. That is not an error.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApproveResult'}
        default:
          description: fingerprint_required, fingerprint_mismatch, fingerprint_moved, definition_unavailable, not_found, corrupt_state, store_busy, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/review-set:
    get:
      operationId: reviewToolSet
      tags: [tools]
      x-gatte-socket: [operator]
      summary: Every pending and changed tool of one backend, and the manifest of that set (`tool review -server`). Feature `tool_review_set`.
      parameters:
        - {name: server, in: query, required: true, schema: {type: string}}
      responses:
        '200':
          description: OK. Nothing waiting is `tools` empty, `manifest` empty and `approvable` false.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolReviewSet'}
        default:
          description: invalid_argument (no server), not_found (no tool observed on it), ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/approve-set:
    post:
      operationId: approveToolSet
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: |
        Approve one backend's whole review set, all of it or nothing, and only
        while it is still the set the manifest names. Writes one
        `(tool approve)` row per tool and one `(tool approve set)` row.
        Feature `tool_review_set`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApproveSetRequest'}
      responses:
        '200':
          description: |
            Approved, or nothing waited any more (`changed` false). Approving
            a set never clears (1.5.0): a sensitive tool of it is approved and
            `usable: false` until cleared one at a time with `clearTool`; the
            warning `sensitive_uncleared` names them.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApproveSetResult'}
        default:
          description: manifest_required, manifest_mismatch, definition_unavailable, not_found, corrupt_state, store_busy, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/revoke:
    post:
      operationId: revokeTool
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: Return an approved tool to pending. Writes `(tool revoke)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ToolRef'}
      responses:
        '200':
          description: Revoked, or already pending (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RevokeResult'}
        default:
          description: changed_not_revocable, not_found, corrupt_state, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/clear:
    post:
      operationId: clearTool
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: |
        Clear an approved sensitive tool for serving at its approved
        fingerprint (1.5.0, feature `tool_clear`, design/adr/0048). Writes
        `(tool clear)`.
      description: |
        The second decision a sensitive tool needs after approval: approval
        says the definition is not poisoned, clearing says this backend may
        act on the team's behalf through this operation. Refused unless some
        role marked `non_read` names the tool in `tools` (`no_non_read_grant`,
        with what covers it instead in `details.callable_by`): a wildcard or a
        `[role.grants]` name never reaches a sensitive tool, so a clearance
        nobody could use is not made. A change to the definition, or a
        revoke, withdraws the clearance with the approval. The gateway asks
        the same role question on every call.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ToolRef'}
      responses:
        '200':
          description: Cleared, or already cleared at exactly this fingerprint (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClearResult'}
        default:
          description: not_sensitive, not_approved, no_non_read_grant, not_found, corrupt_state, store_busy, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /v1/access/block:
    post:
      operationId: blockSubject
      x-gatte-idempotent: true
      tags: [access]
      x-gatte-socket: [operator]
      summary: Block an analyst from the next request on (design/adr/0031). Writes `(access block)`.
      description: The block is placed first and recorded second. If the record fails the answer is still 200 with `recorded` false and warning `audit_write_failed`, and the block stays.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BlockRequest'}
      responses:
        '200':
          description: Blocked, or already blocked (`changed` false). Warning `never_seen` when the trail has no row from this subject.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BlockResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/access/unblock:
    post:
      operationId: unblockSubject
      x-gatte-idempotent: true
      tags: [access]
      x-gatte-socket: [operator]
      summary: Lift a block. Writes `(access unblock)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BlockRequest'}
      responses:
        '200':
          description: Unblocked, or was not blocked (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BlockResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/audit:
    get:
      operationId: listAudit
      tags: [audit]
      x-gatte-socket: [operator]
      summary: The trail, filtered (`audit -json`).
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, default: 200}
          description: Newest N matching records; clamped to 5000.
        - {name: since, in: query, schema: {type: string, format: date-time}}
        - name: subject
          in: query
          schema: {type: string}
          description: Exact match on analyst_identity only, never on analyst_name (design/adr/0037).
        - {name: outcome, in: query, schema: {$ref: '#/components/schemas/AuditOutcome'}}
        - {name: source, in: query, schema: {type: string}, description: Exact source address.}
        - name: until
          in: query
          schema: {type: string, format: date-time}
          description: 1.4.0 (`audit_filters`). Only records BEFORE this time (exclusive; `since` is inclusive). Must be after `since` (`bad_request`).
        - name: tool
          in: query
          schema: {type: string}
          description: 1.4.0 (`audit_filters`). Exact `tool`, e.g. `casemgmt.list_cases` or `(access block)`.
        - name: server
          in: query
          schema: {type: string}
          description: 1.4.0 (`audit_filters`). Exact `target_upstream`. An older backend ignores these three; a front must not send them without the feature.
        - name: before
          in: query
          schema: {type: integer, minimum: 1}
          description: Only records with a `position` lower than this (from `next_before`), to walk the trail backwards.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditPage'}
        default: {$ref: '#/components/responses/Error'}
  /v1/audit/verify:
    post:
      operationId: verifyAudit
      tags: [audit]
      x-gatte-socket: [operator]
      summary: Walk the hash chain and, if given, compare the head (`audit -verify`). Writes nothing.
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VerifyRequest'}
      responses:
        '200':
          description: The verification ran; read `intact` and `head_matches`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VerifyResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/upstreams:
    get:
      operationId: listUpstreams
      tags: [upstreams]
      x-gatte-socket: [operator]
      summary: Registered backends, whether each is signed (`upstream list`) and, with `backend_health`, its health. Maintenance is under `/v1/maintenance`; with `console_manages`, register is `POST` here, deregister `DELETE /v1/upstreams/{name}` and sign is on the accounts socket.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [upstreams]
                properties:
                  upstreams:
                    type: array
                    items: {$ref: '#/components/schemas/Upstream'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: registerUpstream
      x-gatte-idempotent: false
      tags: [upstreams, manage]
      x-gatte-socket: [operator]
      summary: Register a REST API from its OpenAPI document -- the same ingestion, refusals and report as `upstream register -transport http` (1.6.0, `console_manages`). Writes `(upstream register)`. Only http; stdio and oci stay in the terminal.
      description: |
        Not idempotent: a second call answers `upstream_exists`. After an
        `internal` or a dropped connection, `GET /v1/upstreams/{name}` first.
        The entry is NOT signed by this: sign it on the accounts socket.
        Each tool waits in the quarantine. Body at most 24 MiB.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterUpstreamRequest'}
      responses:
        '201':
          description: Registered; the ingestion's report.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RegisterUpstreamResult'}
        default:
          description: feature_disabled, invalid_argument (any refusal of the ingestion, with its message), upstream_exists, payload_too_large, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/upstreams/{name}:
    get:
      operationId: getUpstream
      tags: [upstreams, manage]
      x-gatte-socket: [operator]
      summary: One entry in detail -- for http its descriptor (names only) and each operation with method, path and class -- with its signature state and every tool the quarantine holds under the name (1.6.0, `console_manages`).
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UpstreamDetail'}
        default:
          description: feature_disabled, not_found, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      operationId: deregisterUpstream
      x-gatte-idempotent: true
      tags: [upstreams, manage]
      x-gatte-socket: [operator]
      summary: Deregister a backend, as `upstream deregister` does -- the registry row, its signature, every approval of its tools and its health, all keyed by the name (1.6.0, `console_manages`). Writes `(upstream deregister)`.
      description: |
        The body repeats the name (`confirm`); a front asks the operator to
        type it, because the approvals go with the entry. A name with
        nothing left under it answers `changed: false` and writes no row.
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/GatteFront'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DeregisterRequest'}
      responses:
        '200':
          description: Deregistered, or nothing was there.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeregisterResult'}
        default:
          description: feature_disabled, invalid_argument (confirm), internal (a later step failed; the earlier ones took effect and repeating is safe), ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /v1/secrets:
    get:
      operationId: listSecrets
      tags: [manage]
      x-gatte-socket: [accounts]
      summary: The vault's names -- never a value -- and, for each name a registered entry declares, whether the vault has it (1.6.0, `console_manages`).
      description: |
        The registry is read by a child running as the database's owner
        (root never opens the service account's database); when that
        fails, `registry_read` is false and only the vault's names are
        listed. The vault is decrypted by `sops` from standard input; its
        two files are held to the same ownership check as `config.toml`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SecretList'}
        default: {$ref: '#/components/responses/Error'}
  /v1/secrets/{name}:
    put:
      operationId: setSecret
      x-gatte-idempotent: true
      tags: [manage]
      x-gatte-socket: [accounts]
      summary: Write one vault value (1.6.0, `console_manages`). Writes `(secret set)` with the name, never the value.
      description: |
        `name` is an environment variable name (`[A-Za-z_][A-Za-z0-9_]*`,
        at most 128). The vault is decrypted and encrypted again by `sops`
        on standard input and output, for the file's own age recipients;
        the plaintext never reaches the disk; the new file is decrypted
        once more and compared before it replaces the old one (temporary
        file and rename, owner and mode kept). A vault encrypted for any
        key other than age is refused. A connected backend keeps the old
        value until `POST /v1/upstreams/redial`. Body at most 448 KiB.
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/GatteFront'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SecretValue'}
      responses:
        '200':
          description: Written; `existed` says it replaced a value.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SecretResult'}
        default:
          description: feature_disabled, invalid_argument (name, empty value), payload_too_large, config_unavailable, internal, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      operationId: deleteSecret
      x-gatte-idempotent: true
      tags: [manage]
      x-gatte-socket: [accounts]
      summary: Remove one vault value (1.6.0, `console_manages`). Writes `(secret delete)`; a name the vault does not hold answers `changed` false.
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/GatteFront'}
      responses:
        '200':
          description: Deleted, or not there.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SecretResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/roles:
    get:
      operationId: getRoles
      tags: [manage]
      x-gatte-socket: [accounts]
      summary: The roles file (`roles_file`) as it is on disk (1.6.0, `console_manages`). `feature_disabled` with `details.key` `roles_file` when the roles live in the main file.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RolesText'}
        default: {$ref: '#/components/responses/Error'}
    put:
      operationId: putRoles
      x-gatte-idempotent: true
      tags: [manage]
      x-gatte-socket: [accounts]
      summary: Validate and write the roles file (1.6.0, `console_manages`). Writes `(roles set)` with the text's SHA-256. Does NOT apply it.
      description: |
        The text is validated as the configuration load would read it --
        unknown and repeated keys, and the whole configuration validated
        with these roles -- and a text it would refuse is answered
        `invalid_argument` (`details.field` `text`) with the load's message,
        and not written. A valid text is written atomically (temporary file
        and rename, owner and mode kept). The running gateway applies roles
        on a reload only: a front calls `POST /v1/reload` on the OPERATOR
        socket next. The same text again answers `changed` false. Body at
        most 6 MiB.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RolesRequest'}
      responses:
        '200':
          description: Written (`reload_needed`), or already this text.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RolesResult'}
        default:
          description: feature_disabled, invalid_argument, payload_too_large, config_unavailable, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /v1/serve-requests/{id}:
    get:
      operationId: serveRequest
      tags: [serve]
      x-gatte-socket: [operator]
      summary: Read a request to the gateway process back, e.g. one answered `pending`. Feature `serve_control`.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ServeRequest'}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/people:
    get:
      operationId: people
      tags: [people]
      x-gatte-socket: [operator]
      summary: Roles with their groups and tools, and who the trail has seen.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/People'}
        default: {$ref: '#/components/responses/Error'}
...